refactor: rename Install components to Download
- Rename Install page to Download page
- Move install components to download directory
- Update related imports and references
- Update navigation links from Install to Download
